dspy/pyproject.toml
[project]
name = "dspy-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "datasets>=4.4.1",
    "dspy>=3.0.4",
    "openai>=2.9.0",
    "python-dotenv>=1.0.0",
]
